+++ google/cloud/storage/bucket.py
@@ -52,6 +52,8 @@ from google.cloud.storage.constants import (
 from google.cloud.storage.constants import MULTI_REGIONAL_LEGACY_STORAGE_CLASS
 from google.cloud.storage.constants import MULTI_REGION_LOCATION_TYPE
 from google.cloud.storage.constants import NEARLINE_STORAGE_CLASS
+from google.cloud.storage.constants import PUBLIC_ACCESS_PREVENTION_ENFORCED
+from google.cloud.storage.constants import PUBLIC_ACCESS_PREVENTION_UNSPECIFIED
 from google.cloud.storage.constants import REGIONAL_LEGACY_STORAGE_CLASS
 from google.cloud.storage.constants import REGION_LOCATION_TYPE
 from google.cloud.storage.constants import STANDARD_STORAGE_CLASS
@@ -85,18 +87,6 @@ _LOCATION_SETTER_MESSAGE = (
     "to `Bucket.create`."
 )
 _API_ACCESS_ENDPOINT = "https://storage.googleapis.com"
-
-PUBLIC_ACCESS_PREVENTION_ENFORCED = "enforced"
-"""Enforced public access prevention value.
-
-See: https://cloud.google.com/storage/docs/public-access-prevention
-"""
-
-PUBLIC_ACCESS_PREVENTION_UNSPECIFIED = "unspecified"
-"""Unspecified public access prevention value.
-
-See: https://cloud.google.com/storage/docs/public-access-prevention
-"""
 
 
 def _blobs_page_start(iterator, page, response):
